CMake: add functions for creating md rom targets
md_add_rom takes a desired target name, an md library, the rom head file and the boot assembly file.
From these, it creates a target with the given name (on which usual CMake target commands can be used, such as adding sources), and generates a CMake target called "targetname.rom", which generates a "targetname.bin".