CMake custom function improvements
Simplify specifying extra includes for z80 with md_target_z80_include_directories instead of needing to pass a list variable to a special z80 sources function.

Add a helper for making extra targets, used for z80 and res.

Make copying/padding the rom file a postbuild step of the target, and ensure that the .bin is considered the executable file that was generated rather than the .out elf.