Make the use of newlib a link configuration rather than needing to build md twice
This is basically done by always building md with just the SGDK functions, and only actually using newlib for the end user. Building md with newlib can still be done with the SGDK_BUILD_WITH_NEWLIB CMake configuration option, but this will also force all projects to use newlib regardless of a project's choice.

Various cleanup will follow, as the CMake for all this is a big mess right now. But the win right now is the built source count dropping from ~150 to ~100, as md is finally only built once!